divided ViewFactory in two different Factories
